Move 'managers' package to 'repositories' package
