+++ src/main/java/smida/techtask/repositories/managers/ReportManager.java
@@ -1,4 +1,4 @@
-package smida.techtask.managers;
+package smida.techtask.repositories.managers;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
