+++ src/main/java/smida/techtask/services/impl/CompanyReportServiceImpl.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import smida.techtask.controllers.dto.CompanyReportsDto;
 import smida.techtask.controllers.dto.ReportDto;
-import smida.techtask.managers.CompanyManager;
-import smida.techtask.managers.CompanyReportsManager;
+import smida.techtask.repositories.managers.CompanyManager;
+import smida.techtask.repositories.managers.CompanyReportsManager;
 import smida.techtask.mappers.CompanyReportsMapper;
 import smida.techtask.mappers.ReportMapper;
 import smida.techtask.services.CompanyReportService;
